@@ -2,35 +2,43 @@
 
 namespace App\Command;
 
+use Psr\Container\ContainerInterface;
+use Symfony\Component\Console\Attribute\AsCommand;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
-use Symfony\Component\DependencyInjection\ContainerAwareTrait;
 use Symfony\Component\Process\Process;
 use Symfony\Component\Workflow\Dumper\GraphvizDumper;
 use Symfony\Component\Workflow\Dumper\StateMachineGraphvizDumper;
 use Symfony\Component\Workflow\StateMachine;
 
-class WorkflowUpdateSvgCommand extends Command implements ContainerAwareInterface
+#[AsCommand(
+    name: 'workflow:build:svg',
+    description: 'Updates the SVG files for the workflows',
+)]
+class WorkflowBuildSvgCommand extends Command
 {
-    use ContainerAwareTrait;
+    public function __construct(
+        private readonly ContainerInterface $workflows,
+        private readonly string $projectDir,
+    ) {
+        parent::__construct();
+    }
 
-    protected function configure()
+    protected function configure(): void
     {
         $this
             ->setName('workflow:build:svg')
-            ->setDescription('Build the SVG')
-            ->addArgument('service_name', InputArgument::REQUIRED, 'The service name of the workflow (ex workflow.article)')
+            ->addArgument('name', InputArgument::REQUIRED, 'The workflow name (ex: article)')
         ;
     }
 
     protected function execute(InputInterface $input, OutputInterface $output): int
     {
-        $name = $input->getArgument('service_name');
+        $name = $input->getArgument('name');
 
-        $workflow = $this->container->get($name);
+        $workflow = $this->workflows->get($name);
         $definition = $workflow->getDefinition();
 
         if ($workflow instanceof StateMachine) {
@@ -47,11 +55,9 @@ class WorkflowUpdateSvgCommand extends Command implements ContainerAwareInterfac
 
         $svg = $process->getOutput();
 
-        $svg = preg_replace('/.*<svg/ms', sprintf('<svg class="img-responsive" id="%s"', str_replace('.', '-', $name)), $svg);
+        $svg = preg_replace('/.*<svg/ms', sprintf('<svg class="img-responsive" id="%s"', str_replace('.', '-', (string) $name)), $svg);
 
-        $shortName = explode('.', $name)[1];
-
-        file_put_contents(sprintf('%s/templates/%s/doc.svg.twig', $this->container->getParameter('kernel.project_dir'), $shortName), $svg);
+        file_put_contents(sprintf('%s/templates/%s/doc.svg.twig', $this->projectDir, $name), $svg);
 
         return 0;
     }
