@@ -8,14 +8,14 @@ use Symfony\Component\Security\Http\Event\LogoutEvent;
 
 class LogoutSubscriber implements EventSubscriberInterface
 {
-    public function onLogoutEvent(LogoutEvent $event)
+    public function onLogoutEvent(LogoutEvent $event): void
     {
         $returnTo = $event->getRequest()->query->get('returnTo');
         if (!$returnTo) {
             return;
         }
 
-        $event->setResponse(new RedirectResponse($returnTo));
+        $event->setResponse(new RedirectResponse((string) $returnTo));
     }
 
     public static function getSubscribedEvents(): array
